LPS-197684 Add button to create a rule to RulesEmptyState
